tests/wpt/tests/css/css-overflow/scroll-markers-focus-active-element.html

<!doctype html>
<meta charset="utf-8">
<title>CSS Test: document.activeElement for ::scroll-marker is scroller</title>
<link rel="help"
  href="https://github.com/flackr/carousel/tree/main/scroll-marker#what-is-the-documentactiveelement-of-a-focused-pseudo-element">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  body {
    margin: 0;
  }

  #scroller {
    width: 600px;
    height: 300px;
    overflow: auto;
    scroll-marker-group: before;
    white-space: nowrap;
  }

  #scroller div {
    background: green;
    display: inline-block;
    width: 600px;
    height: 270px;
  }

  #scroller :first-child {
    background: purple;
  }

  #scroller::scroll-marker-group {
    display: flex;
    height: 20px;
    width: 40px;
  }

  #scroller div::scroll-marker {
    content: "";
    width: 100px;
    height: 20px;
    background-color: blue;
    display: inline-block;
  }

  #scroller div:first-of-type::scroll-marker {
    background-color: brown;
  }
</style>
<div id="scroller">
  <div></div>
  <div></div>
</div>
<script>
  promise_test(async t => {
    actions_promise = new test_driver.Actions()
      .pointerMove(15, 15)
      .pointerDown()
      .pointerUp()
      .send();
    await actions_promise;
    assert_equals(document.activeElement, scroller, "document.activeElement for ::scroll-marker is scroller");
  });
</script>
